Fixed typo.

+++ include/arctk/data/histogram/fixed1.hpp
@@ -59,7 +59,7 @@ namespace arc //! arctk namespace
                 //  == INSTANTIATION ==
               public:
                 //  -- Constructors --
-                inline Fixed(double min_, double max_, double size_) noexcept;
+                inline Fixed(double min_, double max_, size_t size_) noexcept;
 
 
                 //  == METHODS ==
@@ -88,7 +88,7 @@ namespace arc //! arctk namespace
              *  @pre    Size_ must be positive.
              */
             template <typename T>
-            inline Fixed<T, 1>::Fixed(const double min_, const double max_, const double size_) noexcept
+            inline Fixed<T, 1>::Fixed(const double min_, const double max_, const size_t size_) noexcept
               : Histogram<T, 1>(min_, max_, size_)
             {
                 assert(min_ < max_);
